--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "blender/blender_autocomplete"]
-	path = blender/blender_autocomplete
-	url = https://github.com/Korchy/blender_autocomplete.git
