+++ scripts/install_addon.py
