+++ scripts/install_addon.sh
