+++ tests/gokart.py
@@ -3,7 +3,7 @@ from pathlib import Path
 sys.path.append(str(Path(__file__).parent.parent.absolute()))
 
 
-from textToBlender import shape, scene, BlenderLength, joint
+from CodeToCADBlenderProvider import shape, scene, BlenderLength, joint
 
 def createFeature(name, feature):
   source = feature["source"]
