+++ tests/revolveExtrude.py
@@ -2,7 +2,7 @@ import sys
 from pathlib import Path
 sys.path.append(str(Path(__file__).parent.parent.absolute()))
 
-from textToBlender import *
+from CodeToCADBlenderProvider import *
 
 curve("centerline").createLine("1")
 curve("donut").createCircle("10cm").translate("-100cm,0,0").revolve("180d", Axis.Y, "centerline")
